Working on Project list and address details screen changes

+++ lib/pages/project/update_address_progress/controller/update_address_progress_controller.dart
@@ -45,7 +45,7 @@ class UpdateAddressProgressController extends GetxController {
     Get.back(result: true);
     Map<String, dynamic> map = {};
     map["id"] = addressDetailsInfo?.id ?? 0;
-    map["progress"] = progress;
+    map["progress"] = progress.toInt();
     map["status"] = determineStatusFromProgress(progress);
 
     isLoading.value = true;
